reorganizing all the folder structure
