+++ src/blog-components/blog-article/BlogArticle.test.tsx
@@ -3,7 +3,7 @@ import { describe, it, expect } from "vitest";
 
 import { BlogArticle } from "./BlogArticle";
 
-import { agileManifestoContent } from "../assets/articles-content/AgileManifesto";
+import { agileManifestoContent } from "../../assets/articles-content/AgileManifesto";
 
 describe("Blog Article Test", () => {
   it("renders a article with header, sections and footer", () => {
